+++ core/src/codecs/Codec.h
@@ -22,8 +22,8 @@
 #include "DeletedDocsFormat.h"
 #include "IdBloomFilterFormat.h"
 #include "IdIndexFormat.h"
+#include "VectorIndexFormat.h"
 #include "VectorsFormat.h"
-#include "VectorsIndexFormat.h"
 
 namespace milvus {
 namespace codec {
@@ -32,6 +32,9 @@ class Codec {
  public:
     virtual VectorsFormatPtr
     GetVectorsFormat() = 0;
+
+    virtual VectorIndexFormatPtr
+    GetVectorIndexFormat() = 0;
 
     virtual DeletedDocsFormatPtr
     GetDeletedDocsFormat() = 0;
@@ -43,9 +46,6 @@ class Codec {
     /*
     virtual AttrsFormat
     GetAttrsFormat() = 0;
-
-    virtual VectorsIndexFormat
-    GetVectorsIndexFormat() = 0;
 
     virtual AttrsIndexFormat
     GetAttrsIndexFormat() = 0;