+++ core/src/codecs/default/DefaultCodec.cpp
@@ -21,6 +21,7 @@
 
 #include "DefaultDeletedDocsFormat.h"
 #include "DefaultIdBloomFilterFormat.h"
+#include "DefaultVectorIndexFormat.h"
 #include "DefaultVectorsFormat.h"
 
 namespace milvus {
@@ -28,6 +29,7 @@ namespace codec {
 
 DefaultCodec::DefaultCodec() {
     vectors_format_ptr_ = std::make_shared<DefaultVectorsFormat>();
+    vector_index_format_ptr_ = std::make_shared<DefaultVectorIndexFormat>();
     deleted_docs_format_ptr_ = std::make_shared<DefaultDeletedDocsFormat>();
     id_bloom_filter_format_ptr_ = std::make_shared<DefaultIdBloomFilterFormat>();
 }
@@ -35,6 +37,11 @@ DefaultCodec::DefaultCodec() {
 VectorsFormatPtr
 DefaultCodec::GetVectorsFormat() {
     return vectors_format_ptr_;
+}
+
+VectorIndexFormatPtr
+DefaultCodec::GetVectorIndexFormat() {
+    return vector_index_format_ptr_;
 }
 
 DeletedDocsFormatPtr