+++ core/src/segment/Types.h
@@ -21,6 +21,7 @@
 
 #include "segment/DeletedDocs.h"
 #include "segment/IdBloomFilter.h"
+#include "segment/VectorIndex.h"
 #include "segment/Vectors.h"
 
 namespace milvus {
@@ -30,6 +31,7 @@ typedef int64_t doc_id_t;
 
 struct Segment {
     VectorsPtr vectors_ptr_ = std::make_shared<Vectors>();
+    VectorIndexPtr vector_index_ptr_ = std::make_shared<VectorIndex>();
     DeletedDocsPtr deleted_docs_ptr_ = nullptr;
     IdBloomFilterPtr id_bloom_filter_ptr_ = nullptr;
 };